+++ src/main/java/com/example/application/data/requests/asset_watchers/CreateAssetWatcherRequest.java
@@ -4,6 +4,8 @@ import com.example.application.entities.common.TransactionType;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public final class CreateAssetWatcherRequest {
@@ -11,8 +13,8 @@ public final class CreateAssetWatcherRequest {
     private Long portfolioId;
     private String assetSymbol;
     private TransactionType transactionType;
-    private double targetPrice;
-    private double targetAmount;
+    private BigDecimal targetPrice;
+    private BigDecimal targetAmount;
     private boolean isCompleted;
 
 }
